requirements.txt
aiogram>=3.12,<4.0
aiosqlite>=0.20.0
apscheduler>=3.10,<4.0
python-dotenv>=1.0.0
